issuer/pyproject.toml
[project]
name = "openposter-issuer"
version = "0.1.0"
description = "OpenPoster issuer (users, creator handles, node + URL registry)"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "sqlalchemy>=2.0",
  "aiosqlite>=0.20",
  "httpx>=0.27",
  "passlib[bcrypt]>=1.7.4",
  "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
]
